+++ ModulePlayer2.h
@@ -42,6 +42,6 @@ public:
 	bool doPunch2;
 	bool doKick;
 
-	p2Point<float> position;
+	p2Point<int> position;
 	unsigned int Health;
 };
